Fixed flatten_abstract_text
If a field is empty, the string comprehension will crash. I originally
just put an if x.text in there, and this worked. As an alternative
method, I tried to use str(). This just made my normal strings get
joined on the space character ' '. Reverted to the if x.text.